+++ .gitignore
@@ -6,7 +6,7 @@ conf/local/**
 !conf/local/.gitkeep
 
 # ignore potentially sensitive credentials files
-conf/**/*credentials*
+conf/local/credentials.yml
 
 # ignore everything in the following folders
 data/**
