+++ .ipython/profile_default/startup/00-kedro-init.py
