+++ docs/source/index.rst
@@ -1,8 +1,8 @@
-.. python_package documentation master file, created by sphinx-quickstart.
+.. dl_tensorflow documentation master file, created by sphinx-quickstart.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to project python_package's API docs!
+Welcome to project dl_tensorflow's API docs!
 =============================================
 
 .. toctree::
