+++ setup.cfg
@@ -8,4 +8,4 @@ known_third_party=kedro
 
 [tool:pytest]
 addopts=--cov-report term-missing
-        --cov src/python_package -ra
+        --cov src/dl_tensorflow -ra
