+++ src/requirements.txt
@@ -6,10 +6,6 @@ jupyter~=1.0
 jupyter_client>=5.1, < 7.0
 jupyterlab==0.31.1
 kedro==0.16.5
-tensorflow-cpu==2.3.1
-pandas==1.1.2 
-numpy==1.19.2
-scikit-learn==0.23.1
 nbstripout==0.3.3
 pytest-cov~=2.5
 pytest-mock>=1.7.1, <2.0
