+++ src/setup.py
@@ -25,7 +25,43 @@
 #
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""wbdc-tf
-"""
 
-__version__ = "0.1"
+from setuptools import find_packages, setup
+
+entry_point = (
+    "dl_tensorflow = dl_tensorflow.run:run_package"
+)
+
+
+# get the dependencies and installs
+with open("requirements.txt", "r", encoding="utf-8") as f:
+    # Make sure we strip all comments and options (e.g "--extra-index-url")
+    # that arise from a modified pip.conf file that configure global options
+    # when running kedro build-reqs
+    requires = []
+    for line in f:
+        req = line.split("#", 1)[0].strip()
+        if req and not req.startswith("--"):
+            requires.append(req)
+
+setup(
+    name="dl_tensorflow",
+    version="0.1",
+    packages=find_packages(exclude=["tests"]),
+    entry_points={"console_scripts": [entry_point]},
+    install_requires=requires,
+    extras_require={
+        "docs": [
+            "sphinx>=1.6.3, <2.0",
+            "sphinx_rtd_theme==0.4.1",
+            "nbsphinx==0.3.4",
+            "nbstripout==0.3.3",
+            "recommonmark==0.5.0",
+            "sphinx-autodoc-typehints==1.6.0",
+            "sphinx_copybutton==0.2.5",
+            "jupyter_client>=5.1.0, <7.0",
+            "tornado>=4.2, <6.0",
+            "ipykernel>=4.8.1, <5.0",
+        ]
+    },
+)
