+++ src/tests/test_run.py
@@ -39,7 +39,7 @@ from pathlib import Path
 
 import pytest
 
-from python_package.run import ProjectContext
+from dl_tensorflow.run import ProjectContext
 
 
 @pytest.fixture
@@ -49,7 +49,7 @@ def project_context():
 
 class TestProjectContext:
     def test_project_name(self, project_context):
-        assert project_context.project_name == "wbdc-tf"
+        assert project_context.project_name == "dl_tensorflow"
 
     def test_project_version(self, project_context):
         assert project_context.project_version == "0.16.5"
