--- wbdc-tf/src/kedro_tutorial/pipelines/data_engineering/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/home/neurlap/.pyenv/versions/kedro/bin/python"
-}
