--- wbdc-tf/src/requirements.txt
@@ -1,17 +1,0 @@
-black==v19.10b0
-flake8>=3.7.9, <4.0
-ipython~=7.0
-isort>=4.3.21, <5.0
-jupyter~=1.0
-jupyter_client>=5.1, < 7.0
-jupyterlab==0.31.1
-kedro==0.16.5
-tensorflow-cpu==2.3.1
-pandas==1.1.2 
-numpy
-scikit-learn==0.23.1
-nbstripout==0.3.3
-pytest-cov~=2.5
-pytest-mock>=1.7.1, <2.0
-pytest~=5.0
-wheel==0.32.2
